Use coin info from other sources e.g bittrex to return values for coins that aren't known

+++ src/CoinBot.Core/ICoinInfo.cs
@@ -37,7 +37,7 @@ namespace CoinBot.Core
         /// <summary>
         ///     The USD price.
         /// </summary>
-        double? PriceUsd { get; }
+        decimal? PriceUsd { get; }
 
         /// <summary>
         ///     The BTC price.
